Refactored steps
